# api-definitions/Repository.openapi+json.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Repository API",
        "version": "v1"
    },
    "paths": {
...
        "/repository/game-servers/{gameServerId}": {
            "get": {
                "tags": [
                    "GameServers"
                ],
                "parameters": [
                    {
                        "name": "gameServerId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "patch": {
                "tags": [
                    "GameServers"
                ],
                "parameters": [
                    {
                        "name": "gameServerId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "delete": {
                "tags": [
                    "GameServers"
                ],
                "parameters": [
                    {
                        "name": "gameServerId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/repository/game-servers": {
            "get": {
                "tags": [
                    "GameServers"
                ],
                "parameters": [
                    {
                        "name": "gameTypes",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "serverIds",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter",
                        "in": "query",
                        "schema": {
                            "$ref": "#/components/schemas/GameServerFilter"
                        }
                    },
                    {
                        "name": "skipEntries",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "takeEntries",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "schema": {
                            "$ref": "#/components/schemas/GameServerOrder"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "post": {
                "tags": [
                    "GameServers"
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
...
        "/repository/game-servers-stats/{serverId}": {
            "get": {
                "tags": [
                    "GameServersStats"
                ],
                "parameters": [
                    {
                        "name": "serverId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "cutoff",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date-time"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "GameServerFilter": {
                "enum": [
                    "LiveStatusEnabled",
                    "ShowOnPortalServerList",
                    "ShowOnBannerServerList"
                ],
                "type": "string",
                "format": "int32"
            },
            "GameServerOrder": {
                "enum": [
                    "BannerServerListPosition",
                    "GameType"
                ],
                "type": "string",
                "format": "int32"
            },
...
        },
        "securitySchemes": {
            "Bearer": {
                "type": "apiKey",
                "description": "",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "Bearer": []
        }
    ]
}